# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ExamsNepal  API",
        "description": "API Endpoints for Forum Questions",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://192.168.1.78:8000/api",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
        "/student/questions": {
            "get": {
                "tags": [
                    "Forum"
                ],
                "summary": "Fetch all questions based on the user's stream",
                "operationId": "5f8975f74f0dd4ec19f65c2eea3847af",
                "responses": {
                    "200": {
                        "description": "List of questions"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/myquestions": {
            "get": {
                "tags": [
                    "Forum"
                ],
                "summary": "Fetch all questions created by me",
                "operationId": "c9e58b3446dbf1f37afac09db1dcf2c6",
                "responses": {
                    "200": {
                        "description": "List of questions created by me"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/addquestion": {
            "post": {
                "tags": [
                    "Forum"
                ],
                "summary": "Add a new question",
                "description": "Adds a new question to the forum for the authenticated user.",
                "operationId": "addQuestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question"
                                ],
                                "properties": {
                                    "question": {
                                        "description": "The question to be added.",
                                        "type": "string",
                                        "example": "What is the best way to improve my coding skills?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Question added successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Question added successfully"
                                        },
                                        "question": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "question": {
                                                    "type": "string",
                                                    "example": "What is the best way to improve my coding skills?"
                                                },
                                                "stream": {
                                                    "type": "string",
                                                    "example": "Computer Science"
                                                },
                                                "deleted": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Student profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Student profile not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "question": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The question field is required"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}